requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Flask-CORS==4.0.0
pdfplumber==0.10.3
beautifulsoup4==4.12.2
requests==2.31.0
reportlab==4.0.7
pydantic==2.8.2
pymysql
spacy
pydantic[email]
scikit-learn
SQLAlchemy>=2.0.0
aiohttp==3.9.1
python-dotenv==1.0.0
